Path registration with dataset QC

+++ 2025-05_trials-extraction/03-re-extract-trials-popeye.py
@@ -1,3 +1,4 @@
+import argparse
 import os
 import pickle
 import logging
@@ -17,6 +18,7 @@ logger = logging.getLogger('ibllib')
 assert correct_version(__version__)
 setenv()
 
+N_JOBS = 48  # Number of processes to use for multiprocessing
 RUN_LIST = PROCESSED.with_stem(f'{REVISION_FPGA}_run_list')
 assert RUN_LIST.exists(), f'Run list file {RUN_LIST} does not exist. Please run 02_get-trials-run-list.py first.'
 
@@ -66,15 +68,22 @@ def process_one_session(tup, processed=None, processed_paths=None):
             status = task.run()
             assert status == 0, 'extraction failure'
             processed_paths[eid] += [popeye_to_sdsc(f) for f in task.outputs if f.name in DATASETS]
+            err.append(None)
         except Exception as ex:
             err.append(str(ex))
             print(("error!!!!!!", err, task.log),)
         finally:
             try:
-                task.cleanUp()  # remove symlinks, etc.
+                task.cleanUp()  # the popeye handler does nothing here
                 # To keep number of files low, remove any other extracted datasets
                 for f in task.outputs:
                     if not f.name in DATASETS:
+                        f.unlink()
+                # remove symlinks immediately -
+                # They won't be registered and there's a file limit on Popeye
+                assert task.data_handler.patch_path.parts[0:4] == task.session_path.parts[0:4]
+                for f in task.session_path.rglob('*.*'):
+                    if f.is_symlink() and f not in task.outputs and f.name not in DATASETS:
                         f.unlink()
             except AssertionError:
                 pass
@@ -96,12 +105,12 @@ def group_by_subject(run_list):
     return groupby(run_list, key=lambda x: x[1].session_parts[1])
 
 
-def process_one_subject(subject, run_list):
+def process_one_subject(subject, run_list, **kwargs):
     """Process all sessions for a given subject."""
     logger.info(f'Processing subject: {subject}')
     out = {}
     for tup in run_list:
-        out.update(process_one_session(tup))
+        out.update(process_one_session(tup, **kwargs))
     return {subject: out}
 
 
@@ -131,24 +140,51 @@ def save_processed(processed, processed_paths):
 
 
 def main():
+    parser = argparse.ArgumentParser(description='Re-extract trials from Popeye sessions.')
+    parser.add_argument('--n-sessions', type=int, default=2,
+                        help='Number of sessions to process (default: 2)')
+    parser.add_argument('--n-subjects', type=int)
+    parsed = parser.parse_args()
+
     with FileLock(RUN_LIST, timeout_action='raise'):
         with open(RUN_LIST, 'rb') as fp:
             run_list = pickle.load(fp)
+    if PROCESSED.exists():
+        with FileLock(PROCESSED, timeout_action='raise'):
+            with open(PROCESSED, 'rb') as fp:
+                _processed = pickle.load(fp)
+            l_before= len(run_list)
+            run_list = list(filter(lambda x: x[0] not in _processed, run_list))
+            logger.info(f'Skipping {l_before - len(run_list)} sessions processed but not registered.')
 
     manager = Manager()
     # Initialize shared dictionaries for processed results and paths
     processed = manager.dict()  # Store processed results
     processed_paths = manager.dict()  # Store processed paths
 
-    f = partial(process_one_session, processed=processed, processed_paths=processed_paths)
-    with Pool(processes=2) as pool:
-        results = pool.map(f, run_list[:2])
-    assert len(processed) == 2
-    assert len(processed_paths) == 2
+    if parsed.n_subjects is not None:
+        n = parsed.n_subjects
+        to_run = []
+        for subject, run_list_ in group_by_subject(run_list):
+            if len(to_run) < n:
+                to_run.append((subject, list(run_list_)))
+            else:
+                break
+        f = partial(process_one_subject, processed=processed, processed_paths=processed_paths)
+    else:
+        n = parsed.n_sessions
+        to_run = run_list[:n]
+        f = partial(process_one_session, processed=processed, processed_paths=processed_paths)
+        
+    with Pool(processes=N_JOBS) as pool:
+        results = pool.map(f, to_run)
     # Save processed paths
     save_processed(dict(processed), dict(processed_paths))
     cmd = f'chmod -R 777 {str(TASKS_DIR)}/Trials*'
     os.system(cmd)
+    for p in (PROCESSED_PATHS, PROCESSED):
+        cmd = f'chmod -R 777 {str(p)}'
+        os.system(cmd)
 
 
 if __name__ == '__main__':
